smbtatorture.c: fix the time measurement

+++ src/torture/smbtatorture.c
@@ -545,9 +545,8 @@ int main(int argc, char *argv[])
 		}
 	}
 	end_time=times(&end_cpu);
-	long double timet = (long double) (end_time - start_time) / clockticks;	
-	printf("Time required for the copying: %lu seconds\n",
-		(unsigned long int) timet);
+	double timet = (double) (end_time - start_time) / (double) clockticks;	
+	printf("Time required for the copying: %f seconds\n", timet);
 	// free allocated memory
 	free(config.user);
 	free(config.workgroup);
